# api-docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Manga Server API",
    "description": "A generic manga website backend API service",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
    "/api/favorites/{mangaId}": {
      "post": {
        "tags": [
          "Favorites"
        ],
        "summary": "Add manga to favorites",
        "operationId": "addFavorite",
        "parameters": [
          {
            "name": "mangaId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Favorites"
        ],
        "summary": "Remove manga from favorites",
        "operationId": "removeFavorite",
        "parameters": [
          {
            "name": "mangaId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          }
        }
      }
    },
    "/api/chapters": {
      "post": {
        "tags": [
          "Chapters"
        ],
        "summary": "Create a new chapter (Admin only)",
        "operationId": "createChapter",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChapterCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseChapterResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/register": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Register a new user",
        "operationId": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseAuthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Login user",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseAuthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/manga/import": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Import a single manga",
        "operationId": "importManga",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MangaImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseMangaResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/manga/batch-import": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Batch import multiple manga",
        "operationId": "batchImportManga",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/MangaImportRequest"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseListMangaResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/manga/{id}": {
      "get": {
        "tags": [
          "Manga"
        ],
        "summary": "Get manga by ID",
        "operationId": "getMangaById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseMangaResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/favorites": {
      "get": {
        "tags": [
          "Favorites"
        ],
        "summary": "Get user's favorite manga",
        "operationId": "getUserFavorites",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageMangaResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/favorites/{mangaId}/status": {
      "get": {
        "tags": [
          "Favorites"
        ],
        "summary": "Check if manga is in favorites",
        "operationId": "isFavorite",
        "parameters": [
          {
            "name": "mangaId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseBoolean"
                }
              }
            }
          }
        }
      }
    },
    "/api/chapters/{id}": {
      "get": {
        "tags": [
          "Chapters"
        ],
        "summary": "Get chapter by ID",
        "operationId": "getChapterById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseChapterResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/chapters/manga/{mangaId}": {
      "get": {
        "tags": [
          "Chapters"
        ],
        "summary": "Get all chapters for a manga",
        "operationId": "getChaptersByMangaId",
        "parameters": [
          {
            "name": "mangaId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseListChapterResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/chapters/manga/{mangaId}/chapter/{chapterNumber}": {
      "get": {
        "tags": [
          "Chapters"
        ],
        "summary": "Get chapter by manga ID and chapter number",
        "operationId": "getChapterByMangaIdAndNumber",
        "parameters": [
          {
            "name": "mangaId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "chapterNumber",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseChapterResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/manga/{id}": {
      "delete": {
        "tags": [
          "Admin"
        ],
        "summary": "Delete a manga",
        "operationId": "deleteManga",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ApiResponseVoid": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object"
          }
        }
      },
      "ChapterCreateRequest": {
        "required": [
          "chapterNumber",
          "mangaId",
          "pageUrls",
          "title"
        ],
        "type": "object",
        "properties": {
          "mangaId": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "chapterNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageUrls": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "ApiResponseChapterResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/ChapterResponse"
          }
        }
      },
      "ChapterResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "mangaId": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "chapterNumber": {
            "type": "integer",
            "format": "int32"
          },
          "viewCount": {
            "type": "integer",
            "format": "int32"
          },
          "pageUrls": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "RegisterRequest": {
        "required": [
          "email",
          "password",
          "username"
        ],
        "type": "object",
        "properties": {
          "username": {
            "maxLength": 50,
            "minLength": 3,
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "maxLength": 2147483647,
            "minLength": 6,
            "type": "string"
          }
        }
      },
      "ApiResponseAuthResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/AuthResponse"
          }
        }
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "role": {
            "type": "string"
          }
        }
      },
      "AuthRequest": {
        "required": [
          "password",
          "username"
        ],
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "ChapterImportRequest": {
        "required": [
          "chapterNumber",
          "pageUrls",
          "title"
        ],
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "chapterNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageUrls": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "MangaImportRequest": {
        "required": [
          "chapters",
          "coverImageUrl",
          "title"
        ],
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "oldName": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "coverImageUrl": {
            "type": "string"
          },
          "detailImages": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "author": {
            "type": "string"
          },
          "slogan": {
            "type": "string"
          },
          "isChoiceness": {
            "type": "boolean"
          },
          "isRecommend": {
            "type": "boolean"
          },
          "isNew": {
            "type": "boolean"
          },
          "period": {
            "type": "string"
          },
          "isPutaway": {
            "type": "boolean"
          },
          "putawayTime": {
            "type": "string"
          },
          "onlineTime": {
            "type": "string"
          },
          "tendency": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "isFinish": {
            "type": "string"
          },
          "sortOrder": {
            "type": "integer",
            "format": "int32"
          },
          "tags": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "labels": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "chapters": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChapterImportRequest"
            }
          }
        }
      },
      "ApiResponseMangaResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/MangaResponse"
          }
        }
      },
      "MangaResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "oldName": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "coverImageUrl": {
            "type": "string"
          },
          "detailImages": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "author": {
            "type": "string"
          },
          "slogan": {
            "type": "string"
          },
          "isChoiceness": {
            "type": "boolean"
          },
          "isRecommend": {
            "type": "boolean"
          },
          "isNew": {
            "type": "boolean"
          },
          "period": {
            "type": "string"
          },
          "isPutaway": {
            "type": "boolean"
          },
          "putawayTime": {
            "type": "string",
            "format": "date-time"
          },
          "onlineTime": {
            "type": "string",
            "format": "date-time"
          },
          "tendency": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "isFinish": {
            "type": "string"
          },
          "sortOrder": {
            "type": "integer",
            "format": "int32"
          },
          "viewCount": {
            "type": "integer",
            "format": "int32"
          },
          "favoriteCount": {
            "type": "integer",
            "format": "int32"
          },
          "chapterCount": {
            "type": "integer",
            "format": "int32"
          },
          "tags": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "labels": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ApiResponseListMangaResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MangaResponse"
            }
          }
        }
      },
      "ApiResponsePageMangaResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/PageMangaResponse"
          }
        }
      },
      "PageMangaResponse": {
        "type": "object",
        "properties": {
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MangaResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableObject"
          },
          "numberOfElements": {
            "type": "integer",
            "format": "int32"
          },
          "empty": {
            "type": "boolean"
          }
        }
      },
      "PageableObject": {
        "type": "object",
        "properties": {
          "offset": {
            "type": "integer",
            "format": "int64"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "paged": {
            "type": "boolean"
          },
          "unpaged": {
            "type": "boolean"
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "SortObject": {
        "type": "object",
        "properties": {
          "empty": {
            "type": "boolean"
          },
          "sorted": {
            "type": "boolean"
          },
          "unsorted": {
            "type": "boolean"
          }
        }
      },
      "ApiResponseBoolean": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "boolean"
          }
        }
      },
      "ApiResponseListChapterResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChapterResponse"
            }
          }
        }
      }
    }
  }
}